chore(next): remove operator, improve do type
