+++ packages/system-next/src/Collections/Immutable/HashMap/core.ts
@@ -1,5 +1,3 @@
-import "../../../Operator"
-
 import type { Refinement } from "../../../Function"
 import { constant, identity, tuple } from "../../../Function"
 import { NoSuchElementException } from "../../../GlobalExceptions"
