+++ packages/system-next/src/Collections/Immutable/List/index.ts
@@ -1,5 +1,3 @@
-import "../../../Operator"
-
 export * from "./api/filterM"
 export * from "./api/mapM"
 export * from "./core"
