+++ packages/system-next/src/Collections/Immutable/Map/index.ts
@@ -1,5 +1,3 @@
-import "../../../Operator"
-
 // codegen:start {preset: barrel, include: ./*.ts, exclude: ./deps.ts}
 export * from "./core"
 // codegen:end
