+++ packages/system-next/src/Collections/Immutable/SortedSet/index.ts
@@ -1,5 +1,3 @@
-import "../../../Operator"
-
 import type { Equal } from "../../../Equal"
 import type { Predicate, Refinement } from "../../../Function"
 import { not } from "../../../Function"
