+++ packages/system-next/src/Equal/index.ts
@@ -1,4 +1,2 @@
-import "../Operator"
-
 export * from "./definition"
 export * from "./operations"
