+++ packages/system-next/src/Function/index.ts
@@ -1,5 +1,3 @@
-import "../Operator"
-
 // codegen:start {preset: barrel, include: ./*.ts}
 export * from "./core"
 export * from "./flow"
