+++ packages/system-next/src/IO/core.ts
@@ -1,5 +1,4 @@
 /* eslint-disable prefer-const */
-import "../Operator"
 
 import * as Tp from "../Collections/Immutable/Tuple"
 import { _A, _U } from "../Effect/definition/commons"
