+++ packages/system-next/src/Ord/index.ts
@@ -1,5 +1,3 @@
-import "../Operator"
-
 export * from "./definition"
 export * from "./operations"
 export * from "./common"
