+++ packages/system-next/src/Ordering/index.ts
@@ -1,4 +1,2 @@
-import "../Operator"
-
 export * from "./definition"
 export * from "./operations"
