+++ packages/system-next/src/Ref/index.ts
@@ -1,5 +1,3 @@
-import "../Operator"
-
 // codegen:start {preset: barrel, include: ./*.ts}
 export * from "./definition"
 // codegen:end
