+++ packages/system-next/src/Support/AtomicBoolean/index.ts
@@ -1,5 +1,3 @@
-import "../../Operator"
-
 import { AtomicReference } from "../AtomicReference"
 
 export class AtomicBoolean extends AtomicReference<boolean> {
