+++ packages/system-next/src/Support/DoublyLinkedList/index.ts
@@ -1,5 +1,3 @@
-import "../../Operator"
-
 export class LinkedListNode<T> {
   public removed = false
   public left: LinkedListNode<T> | undefined
