+++ packages/system-next/src/Support/LinkedList/index.ts
@@ -1,5 +1,3 @@
-import "../../Operator"
-
 export class LinkedListNode<T> {
   constructor(public value: T | null, public next: LinkedListNode<T> | null = null) {}
 }
