+++ packages/system-next/src/Support/OneShot/index.ts
@@ -1,5 +1,3 @@
-import "../../Operator"
-
 export class OneShot<A> {
   private internal: A | undefined = undefined
