+++ packages/system-next/src/Transactional/STM/TryCommit/index.ts
@@ -1,5 +1,3 @@
-import "../../../Operator"
-
 import type * as T from "../../../Effect"
 import type { Journal } from "../Journal"
