+++ packages/system-next/src/Transactional/STM/TxnId/index.ts
@@ -1,5 +1,3 @@
-import "../../../Operator"
-
 import { AtomicNumber } from "../../../Support/AtomicNumber"
 
 export type TxnId = number
