+++ packages/system-next/src/Transactional/STM/index.ts
@@ -1,4 +1,2 @@
-import "../../Operator"
-
 export * from "./core"
 export * from "./gen"
