+++ packages/system-next/src/XPure/index.ts
@@ -1,5 +1,3 @@
-import "../Operator"
-
 /**
  * Core exports
  */
